+++ src/FindAuthority.test.js
@@ -1,18 +1,49 @@
 import {
   render,
   screen,
+  fireEvent,
+  cleanup,
 } from '@testing-library/react';
 
 import FindAuthority from './FindAuthority';
+import { SearchModal } from './components';
 
-const renderFindAuthority = () => render(
-    <FindAuthority
-    />
+jest.mock('./components', () => ({
+  SearchModal: jest.fn(() => <div>SearchModal</div>),
+}));
+
+const renderFindAuthority = (props = {}) => render(
+  <FindAuthority
+    {...props}
+  />,
 );
 
 describe('Given FindAuthority', () => {
-  it('should be rendered', () => {
+  beforeEach(() => {
+    cleanup();
+    jest.clearAllMocks();
     renderFindAuthority();
-    expect(screen.getByText('FindAuthority')).toBeVisible();
+  });
+
+  describe('when component does not get renderCustomTrigger action', () => {
+    it('should render default trigger', () => {
+      expect(screen.getByRole('button', { name: 'ui-plugin-find-authority.linkToMarcAuthorityRecord' })).toBeVisible();
+    });
+  });
+
+  describe('when component get renderCustomTrigger action', () => {
+    it('should handle renderCustomTrigger', () => {
+      const renderCustomTrigger = jest.fn();
+
+      renderFindAuthority({ renderCustomTrigger });
+      expect(renderCustomTrigger).toHaveBeenCalled();
+    });
+  });
+
+  describe('when a user clicks on the default modal trigger', () => {
+    it('should open the modal window', () => {
+      fireEvent.click(screen.getByRole('button', { name: 'ui-plugin-find-authority.linkToMarcAuthorityRecord' }));
+      expect(SearchModal.mock.calls[1][0].open).toBeTruthy();
+    });
   });
 });