+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 ## [0.1.0] (IN PROGRESS)
+
+* Create an empty modal window for find Authorities plugin. Refs UIPFAUTH-7